// Shelf.WebApp/ShelfApp/src/app/shared/models/planogram/planogram.ts
export interface StyleRecursiveObj {
    width: number;
    height: number;
    bottom: number;
    left: number;
    zIndex: number;
    idPogObject: number;
}